+++ packages/hardhat-deploy/src/MidlHardhatEnvironment.ts
@@ -2,6 +2,7 @@ import fs from "node:fs";
 import path from "node:path";
 import {
 	AddressPurpose,
+	AddressType,
 	type BitcoinNetwork,
 	type Config,
 	SignMessageProtocol,
@@ -50,7 +51,6 @@ import {
 import { sendBTCTransactions, waitForTransactionReceipt } from "viem/actions";
 import { type StoreApi, createStore } from "zustand";
 import "~/types/context";
-import { DerivationPath, Wallet } from "~/Wallet";
 
 export class MidlHardhatEnvironment {
 	private readonly store: StoreApi<{
@@ -65,11 +65,8 @@ export class MidlHardhatEnvironment {
 	private readonly confirmationsRequired;
 	private readonly btcConfirmationsRequired;
 
-	private accountIndex = 0;
-
 	private walletClient: WalletClient | undefined;
 
-	public wallet: Wallet;
 	private bitcoinNetwork!: BitcoinNetwork;
 	private chain!: Chain;
 
@@ -94,30 +91,6 @@ export class MidlHardhatEnvironment {
 		if (!fs.existsSync(this.deploymentsPath)) {
 			fs.mkdirSync(this.deploymentsPath);
 		}
-
-		let derivationPath: string;
-
-		const testnetOrMainnet =
-			this.bitcoinNetwork.id === "mainnet" ? "mainnet" : "testnet";
-
-		switch (this.userConfig.derivationPath) {
-			case "leather":
-				derivationPath = DerivationPath.Leather[testnetOrMainnet];
-				break;
-			case "xverse":
-				derivationPath = DerivationPath.Xverse[testnetOrMainnet];
-				break;
-			default:
-				derivationPath =
-					this.userConfig.derivationPath ||
-					DerivationPath.Xverse[testnetOrMainnet];
-		}
-
-		this.wallet = new Wallet(
-			this.userConfig.mnemonic,
-			this.bitcoinNetwork,
-			derivationPath,
-		);
 	}
 
 	private initializeNetwork() {
@@ -176,19 +149,21 @@ export class MidlHardhatEnvironment {
 	}
 
 	public async initialize(accountIndex = 0) {
-		this.accountIndex = accountIndex;
 		this.config = createConfig({
 			networks: [this.bitcoinNetwork],
 			connectors: [
 				keyPairConnector({
-					keyPair: this.wallet.getAccount(this.accountIndex),
+					mnemonic: this.userConfig.mnemonic,
+					paymentAddressType: AddressType.P2WPKH,
+					accountIndex,
 				}),
 			],
+			defaultPurpose: this.userConfig.defaultPurpose,
 			provider: this.userConfig.provider,
 		});
 
 		await connect(this.config, {
-			purposes: [AddressPurpose.Ordinals],
+			purposes: [AddressPurpose.Payment, AddressPurpose.Ordinals],
 		});
 
 		this.store.setState({
@@ -373,8 +348,6 @@ export class MidlHardhatEnvironment {
 
 		const walletClient = await this.getWalletClient();
 
-		const account = getDefaultAccount(this.config);
-
 		if (shouldComplete) {
 			await addCompleteTxIntention(this.config, assetsToWithdraw);
 		}
@@ -416,7 +389,7 @@ export class MidlHardhatEnvironment {
 
 			if (intention.meta?.contractName) {
 				const contractAddress = getContractAddress({
-					from: getEVMAddress(account, this.config.getState().network),
+					from: this.getEVMAddress(),
 					nonce: BigInt(intention.evmTransaction.nonce ?? 0),
 				});
 
@@ -538,12 +511,14 @@ export class MidlHardhatEnvironment {
 			throw new Error("MidlHardhatEnvironment not initialized");
 		}
 
-		const account = getDefaultAccount(this.config);
+		return getEVMAddress(this.getAccount(), this.bitcoinNetwork);
+	}
 
-		if (!account) {
-			throw new Error("No default account set");
+	public getAccount() {
+		if (!this.config) {
+			throw new Error("MidlHardhatEnvironment not initialized");
 		}
 
-		return getEVMAddress(account, this.config.getState().network);
+		return getDefaultAccount(this.config);
 	}
 }