+++ packages/react/src/__tests__/keyPair.ts
@@ -13,3 +13,6 @@ export const getKeyPair = (network: Network = bitcoin.networks.regtest) =>
 		),
 		{ network },
 	);
+
+export const __TEST__MNEMONIC__ =
+	"face spike layer label health knee cry taste carpet found elegant october";