+++ packages/hardhat-deploy/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @midl-xyz/hardhat-deploy
 
+## 2.0.0-next.58
+
+### Patch Changes
+
+- feat(hardhat-deploy): segwit
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@2.0.0-next.58
+  - @midl-xyz/midl-js-react@2.0.0-next.58
+  - @midl-xyz/midl-js-core@2.0.0-next.58
+  - @midl-xyz/midl-js-node@2.0.0-next.58
+
+## 2.0.0-next.57
+
+### Patch Changes
+
+- fix(hardhat-deploy): p2wpkh address purpose
+
+## 2.0.0-next.56
+
+### Patch Changes
+
+- feat(hardhat-deploy): change default purpose to payment
+
 ## 2.0.0-next.54
 
 ### Patch Changes