+++ packages/satoshi-kit/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @midl-xyz/satoshi-kit
 
+## 2.0.0-next.58
+
+### Patch Changes
+
+- feat(hardhat-deploy): segwit
+- Updated dependencies
+  - @midl-xyz/midl-js-react@2.0.0-next.58
+  - @midl-xyz/midl-js-core@2.0.0-next.58
+  - @midl-xyz/midl-js-connectors@2.0.0-next.58
+
 ## 2.0.0-next.54
 
 ### Patch Changes